Move home page to pages
