+++ lib/core/widgets/bottom_nav_bar.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class BottomNavBar extends StatelessWidget {
+  const BottomNavBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return NavigationBar(
+      labelBehavior: NavigationDestinationLabelBehavior.alwaysShow,
+      destinations: const [
+        NavigationDestination(
+          icon: Icon(Icons.home_filled),
+          label: 'Inicio',
+        ),
+        NavigationDestination(
+          icon: Icon(Icons.menu_book),
+          label: 'Cursos',
+        ),
+        NavigationDestination(
+          icon: Icon(Icons.calendar_today),
+          label: 'Pendientes',
+        ),
+        NavigationDestination(
+          icon: Icon(Icons.person),
+          label: 'Perfil',
+        ),
+      ],
+    );
+  }
+}
