+++ lib/core/widgets/main_app_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class MainAppBar extends StatelessWidget {
+class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
   const MainAppBar({
     super.key,
   });
@@ -25,4 +25,7 @@ class MainAppBar extends StatelessWidget {
       backgroundColor: Colors.white,
     );
   }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(56.0);
 }
